Moving utilities, adding more unit tests
